requirements.txt
os
json
requests
python-dotenv
sqlite3

